Small changes to the level selector so I can switch to my desktop

+++ src/Austin/test/LevelManagerTests.cs
@@ -118,14 +118,14 @@ namespace AustinsTests {
 
         [TestCase]
         public void Unit_easyDifficultyTableMaxRound() {
-            levelManager.setDifficulty((int)Difficulty.Easy);
+            levelManager.setDifficulty(Difficulty.Easy);
 
             AssertThat(levelManager.level.maxRound).IsEqual(100);
         }
 
         [TestCase]
         public void Unit_easyDifficultyTableEnemyRanks() {
-            levelManager.setDifficulty((int)Difficulty.Easy);
+            levelManager.setDifficulty(Difficulty.Easy);
 
             Godot.Collections.Array<int> easyRanks = new Godot.Collections.Array<int> {
                 (int)Chicken.Cost.ChickenR1,
@@ -137,7 +137,7 @@ namespace AustinsTests {
 
         [TestCase]
         public void Unit_easyDifficultyTableRoundDifficultyValue() {
-            levelManager.setDifficulty((int)Difficulty.Easy);
+            levelManager.setDifficulty(Difficulty.Easy);
 
             int[] easyRoundDifficulty = new int [levelManager.level.maxRound];
             easyRoundDifficulty[0] = 6;
@@ -150,14 +150,14 @@ namespace AustinsTests {
 
         [TestCase]
         public void Unit_mediumDifficultyTableMaxRound() {
-            levelManager.setDifficulty((int)Difficulty.Medium);
+            levelManager.setDifficulty(Difficulty.Medium);
 
             AssertThat(levelManager.level.maxRound).IsEqual(100);
         }
 
         [TestCase]
         public void Unit_mediumDifficultyTableEnemyRanks() {
-            levelManager.setDifficulty((int)Difficulty.Medium);
+            levelManager.setDifficulty(Difficulty.Medium);
 
             Godot.Collections.Array<int> mediumRanks = new Godot.Collections.Array<int> {
                 (int)Chicken.Cost.ChickenR1,
@@ -170,7 +170,7 @@ namespace AustinsTests {
 
         [TestCase]
         public void Unit_mediumDifficultyTableRoundDifficultyValue() {
-            levelManager.setDifficulty((int)Difficulty.Medium);
+            levelManager.setDifficulty(Difficulty.Medium);
 
             int[] mediumRoundDifficulty = new int [levelManager.level.maxRound];
             mediumRoundDifficulty[0] = 8;
@@ -183,14 +183,14 @@ namespace AustinsTests {
 
         [TestCase]
         public void Unit_hardDifficultyTableMaxRound() {
-            levelManager.setDifficulty((int)Difficulty.Hard);
+            levelManager.setDifficulty(Difficulty.Hard);
 
             AssertThat(levelManager.level.maxRound).IsEqual(100);
         }
 
         [TestCase]
         public void Unit_hardDifficultyTableEnemyRanks() {
-            levelManager.setDifficulty((int)Difficulty.Hard);
+            levelManager.setDifficulty(Difficulty.Hard);
 
             Godot.Collections.Array<int> hardRanks = new Godot.Collections.Array<int> {
                 (int)Chicken.Cost.ChickenR1,
@@ -203,7 +203,7 @@ namespace AustinsTests {
 
         [TestCase]
         public void Unit_hardDifficultyTableRoundDifficultyValue() {
-            levelManager.setDifficulty((int)Difficulty.Hard);
+            levelManager.setDifficulty(Difficulty.Hard);
 
             int[] hardRoundDifficulty = new int [levelManager.level.maxRound];
             hardRoundDifficulty[0] = 15;
@@ -223,21 +223,21 @@ namespace AustinsTests {
 
         [TestCase]
         public void Unit_setDifficultyEasy() {
-            levelManager.setDifficulty((int)Difficulty.Easy);
+            levelManager.setDifficulty(Difficulty.Easy);
 
             AssertThat(levelManager.baseDifficulty).IsEqual(Difficulty.Easy);
         }
 
         [TestCase]
         public void Unit_setDifficultyMedium() {
-            levelManager.setDifficulty((int)Difficulty.Medium);
+            levelManager.setDifficulty(Difficulty.Medium);
 
             AssertThat(levelManager.baseDifficulty).IsEqual(Difficulty.Medium);
         }
 
         [TestCase]
         public void Unit_setDifficultyHard() {
-            levelManager.setDifficulty((int)Difficulty.Hard);
+            levelManager.setDifficulty(Difficulty.Hard);
 
             AssertThat(levelManager.baseDifficulty).IsEqual(Difficulty.Hard);
         }
@@ -245,7 +245,7 @@ namespace AustinsTests {
         [TestCase]
         public void Unit_setMapMeadows() {
             PackedScene meadows = GD.Load<PackedScene>("res://src/Austin/scenes/meadows.tscn");
-            levelManager.setMap((int)AvailableMaps.Meadows);
+            levelManager.setMap(AvailableMaps.Meadows);
 
             AssertThat(levelManager.level.mapScene).IsEqual(meadows);
         }
@@ -253,7 +253,7 @@ namespace AustinsTests {
         [TestCase]
         public void Unit_setMapMultipath() {
             PackedScene multipath = GD.Load<PackedScene>("res://src/Austin/scenes/multipath_map.tscn");
-            levelManager.setMap((int)AvailableMaps.Multipath);
+            levelManager.setMap(AvailableMaps.Multipath);
 
             AssertThat(levelManager.level.mapScene).IsEqual(multipath);
         }
@@ -261,7 +261,7 @@ namespace AustinsTests {
         [TestCase]
         public void Unit_setMapDefault() {
             PackedScene defaultMap = GD.Load<PackedScene>("res://src/Austin/scenes/map.tscn");
-            levelManager.setMap((int)AvailableMaps.Default);
+            levelManager.setMap(AvailableMaps.Default);
 
             AssertThat(levelManager.level.mapScene).IsEqual(defaultMap);
         }
